new tsconfig + rewamped oauth routes and google oauth callback
