+++ app/tsconfig.json
@@ -1,6 +1,8 @@
 {
-	"extends": "fastify-tsconfig",
 	"compilerOptions": {
+		"target": "ESNext",
+		"module": "ESNext",
+		"moduleResolution": "Node",
 		"lib": ["dom", "dom.iterable", "ESNext"],
 		"outDir": "dist",
 		"sourceMap": true,
